requirements.txt
tensorflow
numpy
scikit-learn
joblib
pillow
opencv-python
matplotlib
streamlit
